pyproject.toml
[project]
name = "parser"
version = "0.1.0"
description = "Website checker and parser"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.1.0",
    "google-api-python-client>=2.86.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.2",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.8.5",
    "asyncio>=3.4.3",
    "tqdm>=4.66.1"
]
